+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 Improvements:
 * [OE-100](https://openlmis.atlassian.net/browse/OE-100): Added quantity-unit-toggle to Stock on Hand page
+* [OE-101](https://openlmis.atlassian.net/browse/OE-101): Added quantity-unit-toggle to Stock Card page
 
 2.1.8 / 2025-03-31
 ==================